renamind servercast4 to just servercast
